+++ SoC/accelerator/first_accel/matprod.h
@@ -1,4 +1,4 @@
 #define MAX_MEM_SIZE 1024
 #define MAX_MAT_SIZE MAX_MEM_SIZE/32
 
-void matprod(double *m1, double *m2, double *m3, int N1, int N2, int N3);
+void matprod(float *m1, float *m2, float *m3, int N1, int N2, int N3);
