+++ SoC/accelerator/first_accel/solution1/sim/wrapc_pc/matprod.cpp_pre.cpp.tb.cpp
@@ -11,7 +11,7 @@
 
 
 
-void matprod(double *m1, double *m2, double *m3, int N1, int N2, int N3);
+void matprod(float *m1, float *m2, float *m3, int N1, int N2, int N3);
 # 2 "C:/Users/catia/Rui/MSc_Dissertation/SoC/accelerator/first_accel/matprod.cpp" 2
 
 # 1 "C:/Xilinx/Vitis_HLS/2022.2/tps/win64/msys64/mingw64/x86_64-w64-mingw32/include/stdio.h" 1 3
@@ -1165,7 +1165,7 @@ extern "C" {
 
 
 # 6 "C:/Users/catia/Rui/MSc_Dissertation/SoC/accelerator/first_accel/matprod.cpp"
-void matprod(double *m1, double *m2, double *m3, int N1, int N2, int N3) {
+void matprod(float *m1, float *m2, float *m3, int N1, int N2, int N3) {
 #pragma HLS INTERFACE s_axilite port=return bundle=BUS1
 #pragma HLS INTERFACE s_axilite port=N1 bundle=BUS1
 #pragma HLS INTERFACE s_axilite port=N2 bundle=BUS1
@@ -1175,18 +1175,18 @@ void matprod(double *m1, double *m2, double *m3, int N1, int N2, int N3) {
 #pragma HLS INTERFACE m_axi port = m2 depth=MAX_MAT_SIZE
 #pragma HLS INTERFACE m_axi port = m3 depth=MAX_MAT_SIZE
 
- static double regc=0;
+ static float regc=0;
  int i, j, k;
 
- double m1_buffer[1024];
- double m2_buffer[1024];
- double m3_buffer[1024];
+ float m1_buffer[1024];
+ float m2_buffer[1024];
+ float m3_buffer[1024];
 
- memcpy(m1_buffer, (const double*)m1, N1*N2 * sizeof(double));
- memcpy(m2_buffer, (const double*)m2, N2*N3 * sizeof(double));
+ memcpy(m1_buffer, (const float*)m1, N1*N2 * sizeof(float));
+ memcpy(m2_buffer, (const float*)m2, N2*N3 * sizeof(float));
 
  for (i=0, j=0, k=0; i<N1; ) {
-  double mul = m1_buffer[i*N2+k] * m2_buffer[k*N3+j];
+  float mul = m1_buffer[i*N2+k] * m2_buffer[k*N3+j];
   if (k == 0) regc = mul;
   else regc += mul;
   k++;
@@ -1199,24 +1199,24 @@ void matprod(double *m1, double *m2, double *m3, int N1, int N2, int N3) {
  }
 
 
- memcpy((double*)m3, m3_buffer, N1*N3 * sizeof(double));
+ memcpy((float*)m3, m3_buffer, N1*N3 * sizeof(float));
 }
 #ifndef HLS_FASTSIM
 #ifdef __cplusplus
 extern "C"
 #endif
-void apatb_matprod_ir(double *, double *, double *, int, int, int);
+void apatb_matprod_ir(float *, float *, float *, int, int, int);
 #ifdef __cplusplus
 extern "C"
 #endif
-void matprod_hw_stub(double *m1, double *m2, double *m3, int N1, int N2, int N3){
+void matprod_hw_stub(float *m1, float *m2, float *m3, int N1, int N2, int N3){
 matprod(m1, m2, m3, N1, N2, N3);
 return ;
 }
 #ifdef __cplusplus
 extern "C"
 #endif
-void apatb_matprod_sw(double *m1, double *m2, double *m3, int N1, int N2, int N3){
+void apatb_matprod_sw(float *m1, float *m2, float *m3, int N1, int N2, int N3){
 apatb_matprod_ir(m1, m2, m3, N1, N2, N3);
 return ;
 }
